+++ src/gc.c
@@ -2413,8 +2413,7 @@ mark: {
             }
             const jl_datatype_layout_t *layout = jl_task_type->layout;
             assert(layout->fielddesc_type == 0);
-            size_t nfields = layout->nfields;
-            assert(nfields > 0);
+            assert(layout->nfields > 0);
             uint32_t npointers = layout->npointers;
             obj8_begin = (uint8_t*)jl_dt_layout_ptrs(layout);
             obj8_end = obj8_begin + npointers;
@@ -2654,6 +2653,11 @@ JL_DLLEXPORT int64_t jl_gc_diff_total_bytes(void)
 void jl_gc_sync_total_bytes(void)
 {
     jl_gc_get_total_bytes(&last_gc_total_bytes);
+}
+
+JL_DLLEXPORT int64_t jl_gc_live_bytes(void)
+{
+    return live_bytes;
 }
 
 static void jl_gc_premark(jl_ptls_t ptls2)